SDevice::time_ is now float type instead of int type

+++ oneflow/core/schedule/schedule.h
@@ -30,7 +30,7 @@ class Schedule {
  public:
   explicit Schedule(Session* session) : session_(session) {}
 
-  inline const std::unordered_map<TaskInstance*, std::pair<int32_t, int32_t>>&
+  inline const std::unordered_map<TaskInstance*, std::pair<float, float>>&
   instance2ended_at() const {
     return instance2ended_at_;
   }
@@ -44,7 +44,7 @@ class Schedule {
   end_time_gap_to_loss() const {
     return end_time_gap_to_loss_;
   };
-  inline const std::unordered_map<SDevice*, int32_t>& device2ended_at() const {
+  inline const std::unordered_map<SDevice*, float>& device2ended_at() const {
     return device2ended_at_;
   }
   inline const std::unordered_map<STask*, float>& node2interval() const {
@@ -69,7 +69,7 @@ class Schedule {
   }
 
   //	setter
-  inline std::unordered_map<TaskInstance*, std::pair<int32_t, int32_t>>&
+  inline std::unordered_map<TaskInstance*, std::pair<float, float>>&
   mut_instance2ended_at() {
     return instance2ended_at_;
   }
@@ -81,7 +81,7 @@ class Schedule {
   mut_end_time_gap_to_loss() {
     return end_time_gap_to_loss_;
   };
-  inline std::unordered_map<SDevice*, int32_t>& mut_device2ended_at() {
+  inline std::unordered_map<SDevice*, float>& mut_device2ended_at() {
     return device2ended_at_;
   }
   inline std::unordered_map<STask*, float>& mut_node2interval() {
@@ -99,13 +99,12 @@ class Schedule {
 
  protected:
   Session* session_;
-  std::unordered_map<TaskInstance*, std::pair<int32_t, int32_t>>
-      instance2ended_at_;
+  std::unordered_map<TaskInstance*, std::pair<float, float>> instance2ended_at_;
   std::unordered_map<TaskInstance*, std::unordered_map<STask*, float>>
       start_time_gap_to_loss_;
   std::unordered_map<TaskInstance*, std::unordered_map<STask*, float>>
       end_time_gap_to_loss_;
-  std::unordered_map<SDevice*, int32_t> device2ended_at_;
+  std::unordered_map<SDevice*, float> device2ended_at_;
   std::unordered_map<STask*, float> node2interval_;
   float max_interval_ = 0.0;
   std::unordered_map<SRegstDesc*, float> regst_desc2duration_;
